requirements.txt
# Database
alembic>=1.13.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# For async support (optional)
# asyncpg>=0.29.0

# Environment variables
python-dotenv>=1.0.0

# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Email (optional)
# sendgrid>=6.10.0
# boto3>=1.28.0  # For AWS SES

# AI Services (optional)
# openai>=1.0.0
# anthropic>=0.7.0

# Redis (optional)
# redis>=5.0.0

# Utilities
python-multipart>=0.0.6
email-validator>=2.0.0

